refactor: 딥링크 결정 로직 리팩터링
